Backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
chromadb==0.4.24
groq==0.11.0
httpx>=0.24,<0.28
python-dotenv==1.0.1
sentence-transformers==3.3.1
pydantic>=2.12.2,<3
pypdf==5.1.0
numpy<2
openai-agents>=0.11.1
beautifulsoup4>=4.12.0